fix: 입장 안내 이메일 전송 시점 수정
